requirements.txt
Cython==0.29.32
fairseq==0.12.2
gym==0.21.0
gymnasium==0.28.1
h5py==2.10.0
lescode==0.2.13
matplotlib==3.5.3
matplotlib-inline==0.1.6
networkx==2.6.3
numba==0.56.4
numpy==1.21.5
numpydoc==1.4.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
pandas==1.3.5
pettingzoo==1.23.1
pip==22.2.2
PyYAML==6.0
scipy==1.7.3
seaborn==0.11.2
stable-baselines3==1.7.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tianshou==0.5.0
torch-cluster==1.6.1
torch-geometric==2.2.0
torch-scatter==2.1.1
torch-sparse==0.6.17
torch-spline-conv==1.2.2
tornado==6.2
tsplib95==0.7.1
typer==0.7.0
